add config export to init
